+++ pkg/model/user.go
@@ -6,7 +6,7 @@ import (
 
 //User model
 type User struct {
-	ID        uint      `json:"id" gorm:"primary_key"`
+	ID        uint64    `json:"id" gorm:"primary_key"`
 	Name      string    `json:"name"`
 	Document  string    `json:"document"`
 	Email     string    `json:"email" gorm:"type:varchar(100);unique_index"`
